fix: respect query_parallelism for IVF partition prepare
